Разбивает код на отдельные функции
