+++ js/main.js
@@ -1,4 +1,4 @@
 import {renderPhotos} from './pictures.js';
+
 renderPhotos();
-import './bigPictures.js';
